requirements.txt
Django==5.2.13
djangorestframework==3.17.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
PyJWT==2.12.1
sqlparse==0.5.5
asgiref==3.11.1
